lektor.types: remove cyclic imports
