+++ lektor/types/primitives.py
@@ -7,7 +7,7 @@ from pytz import FixedOffset
 
 from lektor.environment import PRIMARY_ALT
 from lektor.i18n import get_i18n_block
-from lektor.types import Type
+from lektor.types.base import Type
 from lektor.utils import bool_from_string
